11-101: adding separate routes
